Fix style side effect race
Setting style on a parent dom node can fail if the two same style instructions
succeeds to each other. i.e. set bg-color on body and navigate to a page setting
the same bg-color on body. Because cancellation is async, the first style
instruction will remove the style on unmount AFTER the second style instruction
has been mounted. `e/apply` is the culprit in the current case, because we
navigate to the same page on link click (with different route params), and
e/apply remounts the conditional branch on arity change.

Idiom: styles should be scoped to a component – don't alter style outside a
component. Or use `css/scoped-style`.