HFQL understands java.util.Enumeration and j.u.Iterator natively
Note the example has slightly changed.
It is expected and correct for `(hfql [.entries] xs)` to return
`{'.entries #object[Iterator 0x123456]}`.
Turning the Iterator object into a seq can be damaging to user's API and should
not be done automatically for rendering convenience. A navigator UI on the other
hand would seq it automatically.
`(hfql [{.entries iterator-seq}] xs)` is not an anti-pattern.